+++ src/main/java/com/chat/app/repositories/base/MessageRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 
-public interface MessageRepository extends JpaRepository<Message, Integer> {
+public interface MessageRepository extends JpaRepository<Message, Long> {
     @Query(value="from Message where receiverId = :userId and session_date = :lastCheckDate and time > :lastCheckTime or receiverId = :userId and session_date > :lastCheckDate")
-    List<Message> findMostRecentMessages(@Param("userId") int userId, @Param("lastCheckDate") LocalDate lastCheckDate, @Param("lastCheckTime") LocalTime lastCheckTime);
+    List<Message> findMostRecentMessages(@Param("userId") long userId, @Param("lastCheckDate") LocalDate lastCheckDate, @Param("lastCheckTime") LocalTime lastCheckTime);
 }
