+++ src/main/java/com/chat/app/repositories/base/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<UserModel, Integer> {
+public interface UserRepository extends JpaRepository<UserModel, Long> {
     List<UserModel> findAll();
 
     UserModel findByUsername(String username);
